Fixing error
Getting an error on the bike and run classes that the distance is parsed as an int. I think it can come down as an int or a double so I am just converting any possible ints to doubles on the parse which will work for both ints and doubles.